+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* Fix in `tbl_uvregression()` for the `formula` argument when it includes a hard-coded column name, e.g. `formula='{y} ~ {x} + grade'`. The hard-coded variable name is now removed from the `include` argument. This was a regression introduced in the v2.0.0 release. (#1886)
+
 # gtsummary 2.0.0
 
 ### New Features